feat(workspace): rename generator to init
