+++ packages/workspace/generators.json
@@ -3,9 +3,9 @@
     "name": "workspace",
     "version": "0.0.1",
     "generators": {
-        "install": {
-            "factory": "./src/generators/install/generator",
-            "schema": "./src/generators/install/schema.json",
+        "init": {
+            "factory": "./src/generators/init/generator",
+            "schema": "./src/generators/init/schema.json",
             "description": "Configure Workspace for Stacks"
         }
     }
