+++ packages/workspace/src/generators/init/files/tsconfig/tsconfig.base.json__template__
