+++ packages/workspace/src/generators/init/schema.d.ts
@@ -4,5 +4,4 @@ export interface InstallGeneratorSchema {
     eslint?: boolean;
     husky?: boolean;
     commitizen?: boolean;
-    skipFormat?: boolean;
 }
