+++ packages/workspace/src/generators/init/utils/constants.ts
