+++ packages/workspace/src/generators/init/utils/eslint.ts
